reorganise util into its own directory
