+++ src/inspect_ai/model/_providers/util/__init__.py
@@ -1,0 +1,22 @@
+from .chatapi import (
+    chat_api_input,
+    chat_api_request,
+    is_chat_api_rate_limit,
+    llama31_chat_api_input,
+    llama31_parse_tool_call,
+    llama31_tools_message,
+)
+from .tools import parse_tool_call
+from .util import as_stop_reason, model_base_url
+
+__all__ = [
+    "as_stop_reason",
+    "chat_api_request",
+    "chat_api_input",
+    "is_chat_api_rate_limit",
+    "llama31_chat_api_input",
+    "llama31_parse_tool_call",
+    "llama31_tools_message",
+    "model_base_url",
+    "parse_tool_call",
+]
