+++ src/inspect_ai/model/_providers/util/util.py
@@ -1,0 +1,36 @@
+import os
+from logging import getLogger
+
+from ..._model_output import StopReason
+
+logger = getLogger(__name__)
+
+
+def as_stop_reason(reason: str | None) -> StopReason:
+    """Encode common reason strings into standard StopReason."""
+    match reason:
+        case "stop" | "eos":
+            return "stop"
+        case "length" | "content_filter":
+            return reason
+        case "model_length":
+            return "length"
+        case "tool_calls" | "function_call":
+            return "tool_calls"
+        case _:
+            return "unknown"
+
+
+def model_base_url(base_url: str | None, env_vars: str | list[str]) -> str | None:
+    if base_url:
+        return base_url
+
+    if isinstance(env_vars, str):
+        env_vars = [env_vars]
+
+    for env_var in env_vars:
+        base_url = os.getenv(env_var, None)
+        if base_url:
+            return base_url
+
+    return os.getenv("INSPECT_EVAL_MODEL_BASE_URL", None)
